+++ NEWS.md
@@ -57,6 +57,9 @@
 
 * `release()` now shows file size before submitting to CRAN (#683, @krlmlr).
 
+* function `complete()` helps complete a session by checking if NEWS(.md),
+  version, and date are up-to-date
+
 * Previously, `devtools` ran all external R processes with `R --vanilla`.
   Now it only suppresses user profiles, and constructs a custom `.Rprofile` to
   override the default.  Currently, this `.Rprofile` sets up the `repos` option.